Rename the module
